fix: show warnings on console (#3225)
Show warnings and errors on apps as well as logs

Also moved a struct that was in the middle of a method into its own file
