+++ applications/tari_base_node/Cargo.toml
@@ -27,7 +27,6 @@ chrono = "0.4"
 config = { version = "0.9.3" }
 futures = { version = "^0.3.1", default-features = false, features = ["alloc"]}
 log = { version = "0.4.8", features = ["std"] }
-log4rs = { version = "0.8.3", features = ["toml_format", "rolling_file_appender", "compound_policy", "size_trigger", "fixed_window_roller"] }
 regex = "1"
 rustyline = "6.0"
 rustyline-derive = "0.3"
@@ -40,3 +39,5 @@ tonic = "0.2"
 [features]
 avx2 = ["tari_core/avx2", "tari_crypto/avx2", "tari_p2p/avx2", "tari_wallet/avx2", "tari_comms/avx2", "tari_comms_dht/avx2"]
 safe = []
+
+
