+++ base_layer/core/src/base_node/state_machine_service/states/block_sync.rs
@@ -99,7 +99,7 @@ impl BlockSync {
                 StateEvent::BlocksSynchronized
             },
             Err(err) => {
-                debug!(target: LOG_TARGET, "Block sync failed: {}", err);
+                warn!(target: LOG_TARGET, "Block sync failed: {}", err);
                 StateEvent::BlockSyncFailed
             },
         }
