+++ base_layer/wallet_ffi/Cargo.toml
@@ -23,7 +23,7 @@ rand = "0.8"
 chrono = { version = "0.4.6", features = ["serde"]}
 thiserror = "1.0.20"
 log = "0.4.6"
-log4rs = {version = "0.8.3", features = ["console_appender", "file_appender", "file", "yaml_format"]}
+log4rs = {version = "1.0.0", features = ["console_appender", "file_appender", "yaml_format"]}
 
 [dependencies.tari_core]
 path = "../../base_layer/core"
