+++ common/logging/log4rs_sample_base_node.yml
@@ -14,8 +14,13 @@ appenders:
   # An appender named "stdout" that writes to stdout
   stdout:
     kind: console
+
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {h({l}):5} {m}{n}"
+      pattern: "{d(%H:%M)} {h({l}):5} {m}{n}"
+    filters:
+      -
+        kind: threshold
+        level: warn
 
   # An appender named "network" that writes to a file with a custom pattern encoder
   network:
@@ -32,7 +37,7 @@ appenders:
         count: 5
         pattern: "log/base-node/network.{}.log"
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {l:5} {m}{n}"
+      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] [Thread:{I}] {l:5} {m}{n}"
 
   # An appender named "base_layer" that writes to a file with a custom pattern encoder
   base_layer:
@@ -49,7 +54,7 @@ appenders:
         count: 5
         pattern: "log/base-node/base_layer.{}.log"
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {l:5} {m}{n}"
+      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] [Thread:{I}] {l:5} {m}{n}"
 
   # An appender named "other" that writes to a file with a custom pattern encoder
   other:
@@ -66,13 +71,13 @@ appenders:
         count: 5
         pattern: "log/base-node/other.{}.log"
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {l:5} {m}{n}"
+      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] [Thread:{I}] {l:5} {m}{n}"
 
 # Set the default logging level to "info"
 root:
-  level: info
+  level: warn
   appenders:
-    - base_layer
+    - stdout
 
 loggers:
   # Route log events common to every application to all appenders
@@ -89,70 +94,61 @@ loggers:
     level: info
     appenders:
       - base_layer
-    additive: false
   tari:
     level: info
     appenders:
       - base_layer
-    additive: false
 
   # Route log events sent to the "wallet" logger to the "base_layer" appender
   wallet:
     level: info
     appenders:
       - base_layer
-    additive: false
   # Route log events sent to the "comms" logger to the "network" appender
   comms:
     level: info
     appenders:
       - network
-    additive: false
   # Route log events sent to the "p2p" logger to the "network" appender
   p2p:
     level: info
     appenders:
       - network
-    additive: false
-   # Route log events sent to the "yamux" logger to the "network" appender
+
+    # Route log events sent to the "yamux" logger to the "network" appender
   yamux:
     level: info
     appenders:
       - network
-    additive: false
   # Route log events sent to the "mio" logger to the "network" appender
   mio:
     level: error
     appenders:
       - network
-    additive: false
   # Route log events sent to the "rustyline" logger to the "other" appender
   rustyline:
     level: error
     appenders:
       - other
     additive: false
+
   # Route log events sent to the "tokio_util" logger to the "other" appender
   tokio_util:
     level: error
     appenders:
       - other
-    additive: false
   # Route PGP log events
   pgp:
-      level: warn
-      appenders:
-        - other
-      additive: false
+    level: warn
+    appenders:
+      - other
   # Route log events sent to the "tari_mm_proxy" logger to the "base_layer" appender
   tari_mm_proxy:
     level: info
     appenders:
       - base_layer
-    additive: false
   # Route log events sent to the "stress_test" logger to the "base_layer" appender
   stress_test:
     level: info
     appenders:
       - base_layer
-    additive: false
