+++ common/logging/log4rs_sample_wallet.yml
@@ -16,6 +16,9 @@ appenders:
     kind: console
     encoder:
       pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {h({l}):5} {m}{n}"
+    filters:
+      - kind: threshold
+        level: warn
 
   # An appender named "network" that writes to a file with a custom pattern encoder
   network:
@@ -32,7 +35,7 @@ appenders:
         count: 5
         pattern: "log/wallet/network.{}.log"
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {l:5} {m}{n}"
+      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] [Thread:{I}] {l:5} {m}{n}"
 
   # An appender named "base_layer" that writes to a file with a custom pattern encoder
   base_layer:
@@ -49,7 +52,7 @@ appenders:
         count: 5
         pattern: "log/wallet/base_layer.{}.log"
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {l:5} {m}{n}"
+      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] [Thread:{I}] {l:5} {m}{n}"
 
        # An appender named "base_layer" that writes to a file with a custom pattern encoder
   other:
@@ -66,11 +69,11 @@ appenders:
         count: 5
         pattern: "log/wallet/other.{}.log"
     encoder:
-      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] {l:5} {m}{n}"
+      pattern: "{d(%Y-%m-%d %H:%M:%S.%f)} [{t}] [Thread:{I}] {l:5} {m}{n}"
 
 # Set the default logging level to "warn" and attach the "stdout" appender to the root
 root:
-  level: warn
+  level: info
   appenders:
     - stdout
 
@@ -88,45 +91,37 @@ loggers:
     level: info
     appenders:
       - base_layer
-    additive: false
   # Route log events sent to the "wallet" logger to the "base_layer" appender
   wallet:
     level: info
     appenders:
       - base_layer
-    additive: false
   # Route log events sent to the "comms" logger to the "network" appender
   comms:
     level: info
     appenders:
       - network
-    additive: false
   p2p:
     level: info
     appenders:
       - network
-    additive: false
     # Route log events sent to the "p2p" logger to the "network" appender
   yamux:
     level: info
     appenders:
       - network
-    additive: false
  # Route log events sent to the "mio" logger to the "other" appender
   mio:
     level: error
     appenders:
       - network
-    additive: false
   # Route log events sent to the "rustyline" logger to the "other" appender
   rustyline:
     level: error
     appenders:
       - other
-    additive: false
   # Route log events sent to the "tokio_util" logger to the "other" appender
   tokio_util:
     level: error
     appenders:
       - other
-    additive: false
